Backed out changeset d2f081b5b95f
Make some Dao functions suspendable. This cause some issues with database
transactions.

A db transaction requires the primary connection in order to
write to the db and this connection is acquired by the thread executing
`db.beginTransaction()`.

A suspendable Dao function will switch context to
`db.queryExecutor.asCoroutineDispatcher()` which by default is a pool of 2 IO
threads. If the suspendable function needs to write to the db, it will try to
acquire the primary connection which is hold by the thread executing the
transaction, thus a deadlock as this thread is waiting for the completion of
the suspendable functions to end the transaction.

This manifests with the following repeating log:

```
12-11 16:18:14.797 18260 18336 W SQLiteConnectionPool: The connection pool for database 'room_articles.db'
has been unable to grant a connection to thread 45671 (arch_disk_io_1) with flags 0x2 for 30.000002 seconds.
12-11 16:18:14.797 18260 18336 W SQLiteConnectionPool: Connections: 0 active, 1 idle, 1 available.
12-11 16:18:44.800 18260 18336 W SQLiteConnectionPool: The connection pool for database 'room_articles.db'
has been unable to grant a connection to thread 45671 (arch_disk_io_1) with flags 0x2 for 60.003002 seconds.
12-11 16:18:44.800 18260 18336 W SQLiteConnectionPool: Connections: 0 active, 1 idle, 1 available.
```