app: TinyrssAccountTokenRetriever don't spam invalidateToken
As the default TinyTinyRss server as a throttling limitation on login, we
don't want to have multiple threads invalidating the token and forcing us to
make multiple login requests in a row.
To do that we make TinyrssAccountTokenRetriever threadsafe by making each
request run in a queue, and let clients wait for result.